requirements.txt
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.85.0
